Extract resources info collecting into EmbeddedResourceExporter
